requirements.txt
flask==2.3.3
pillow==10.0.0
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
transformers==4.31.0
requests==2.31.0
python-dotenv==1.0.0
flask-wtf==1.1.1 